sb-119 update initialization logic
