+++ packages/core/src/components/slide-controls/slide-controls.scss
@@ -54,6 +54,7 @@
     min-width: 14rem;
     justify-content: space-between;
     align-items: center;
+    flex-direction: row-reverse;
 }
 
 .slide-controls__host-link {
