+++ packages/core/src/providers/socket/provider.tsx
@@ -3,12 +3,16 @@
 import { useEffect, useState } from "react";
 import { io, type Socket } from "socket.io-client";
 
+import { CONFIGURED_SERVER } from "@/lib/settings";
+
 import { SocketContext } from "./context";
 
 export const SocketProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
     const [socket, setSocket] = useState<Socket | null>(null);
 
     useEffect(() => {
+        if (!CONFIGURED_SERVER) return;
+
         // Initialize socket connection to our external Socket.io server
         const socketInstance = io(process.env.NEXT_PUBLIC_SERVER_URL || window.location.origin);
