+++ packages/server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@slidebook/server",
-  "version": "0.1.14",
+  "version": "0.1.15",
   "private": true,
   "bin": {
     "slidebook-server": "./server.js"
